refactor: move enable subcommand into own module
